Real struggle
It took a lot of time for me to solve this problem. Was annoyed in
between because I could not get the right answer because of my stupid
mistakes. I was able to get part 1 right but couldn't think of how to
move in each direction up until the bounds were hit. Was pretty simple
once I looked it up in YouTube. GG. The problems are getting harder and
harder. For sure I won't be able to keep up.